feat(cli): refactor login - remove --api-key, add interactive org/project picker
- Remove --api-key, --org-id, --project-id flags from login and init
- Add -y flag to login: skip org/project picker, use session defaults
- Default (no -y): interactive org/project picker after login
- Extract runOrgProjectSelection effect, reuse in orgs switch
- Show init/switch hints after picker when -y not passed
- Move 'already logged in' hint to outro

Made-with: Cursor
