+++ ts/packages/cli/src/commands/login.cmd.ts
@@ -9,25 +9,17 @@ import {
 } from 'src/services/composio-clients';
 import { ComposioUserContext } from 'src/services/user-context';
 import { TerminalUI } from 'src/services/terminal-ui';
+import { runOrgProjectSelection } from 'src/effects/select-org-project';
 
 export const noBrowser = Options.boolean('no-browser').pipe(
   Options.withDefault(false),
   Options.withDescription('Login without browser interaction')
 );
 
-const apiKeyOpt = Options.text('api-key').pipe(
-  Options.optional,
-  Options.withDescription('API key for non-interactive login (agents/CI)')
-);
-
-const orgIdOpt = Options.text('org-id').pipe(
-  Options.optional,
-  Options.withDescription('Organization ID for non-interactive login')
-);
-
-const projectIdOpt = Options.text('project-id').pipe(
-  Options.optional,
-  Options.withDescription('Project ID for non-interactive login')
+const yesOpt = Options.boolean('yes').pipe(
+  Options.withAlias('y'),
+  Options.withDefault(false),
+  Options.withDescription('Skip org/project picker; use session defaults')
 );
 
 /**
@@ -49,6 +41,8 @@ const storeCredentials = (params: {
    * When false, all session/info errors are non-fatal (browser login).
    */
   strictVerification: boolean;
+  /** When true, skip the init/switch hints and outro (shown later after org/project picker). */
+  skipHints?: boolean;
 }) =>
   Effect.gen(function* () {
     const ui = yield* TerminalUI;
@@ -61,6 +55,7 @@ const storeCredentials = (params: {
       initialProjectId,
       fallbackEmail,
       strictVerification,
+      skipHints = false,
     } = params;
 
     // Call session/info to enrich the login with org/project metadata.
@@ -116,10 +111,14 @@ const storeCredentials = (params: {
 
     const email = sessionInfo?.org_member.email || fallbackEmail || undefined;
     yield* ui.log.success(email ? `Logged in as ${email}` : 'Logged in successfully');
-    yield* ui.log.info('Run `composio init` in your project directory to set up project context.');
-    yield* ui.log.info(
-      'To switch your default global org/project later, run `composio orgs switch`.'
-    );
+    if (!skipHints) {
+      yield* ui.log.info(
+        'Run `composio init` in your project directory to set up project context.'
+      );
+      yield* ui.log.info(
+        'To switch your default global org/project later, run `composio orgs switch`.'
+      );
+    }
 
     // Emit structured JSON for piped/scripted consumption (agent-native)
     yield* ui.output(
@@ -132,7 +131,9 @@ const storeCredentials = (params: {
       })
     );
 
-    yield* ui.outro("You're all set!");
+    if (!skipHints) {
+      yield* ui.outro("You're all set!");
+    }
   });
 
 /**
@@ -149,6 +150,8 @@ export const browserLogin = (params: {
   scope: 'user' | 'project';
   /** When true, don't open browser — just show the URL. */
   noBrowser: boolean;
+  /** When true (login only), skip org/project picker and use session defaults. When false, prompt for org/project. */
+  skipOrgProjectPicker?: boolean;
 }) =>
   Effect.gen(function* () {
     const ui = yield* TerminalUI;
@@ -231,86 +234,70 @@ export const browserLogin = (params: {
       initialProjectId: xProjectId,
       fallbackEmail: linkedSession.account.email,
       strictVerification: false,
+      skipHints: params.scope === 'user' && !params.skipOrgProjectPicker,
     });
+
+    if (params.scope === 'user' && !params.skipOrgProjectPicker) {
+      const result = yield* runOrgProjectSelection({
+        apiKey: uakApiKey,
+        baseURL: ctx.data.baseURL,
+      });
+      if (result) {
+        const sessionUserId = uakSessionInfo.org_member.user_id ?? uakSessionInfo.org_member.id;
+        const testUserId = sessionUserId ? `pg-test-${sessionUserId}` : undefined;
+        yield* ctx.login(
+          uakApiKey,
+          result.org.id,
+          result.project.id,
+          testUserId ?? Option.getOrUndefined(ctx.data.testUserId)
+        );
+        yield* ui.log.success(
+          `Default org/project set to "${result.org.name}" / "${result.project.name}".`
+        );
+      }
+      yield* ui.log.info(
+        'Run `composio init` in your project directory to set up project context.'
+      );
+      yield* ui.log.info(
+        'To switch your default global org/project later, run `composio orgs switch`.'
+      );
+      yield* ui.outro("You're all set!");
+    }
   });
 
 /**
  * CLI command to login using Composio's CLI session APIs.
  *
- * Supports two modes:
- * 1. Browser-based: Opens browser for OAuth flow (default)
- * 2. Non-interactive: Accepts --api-key, --org-id, --project-id flags for agents/CI
+ * Browser-based: Opens browser for OAuth flow (default).
+ * Use --no-browser to skip auto-opening the browser and print the URL instead.
+ * Use -y to skip org/project picker and use session defaults.
  *
  * @example
  * ```bash
  * composio login
  * composio login --no-browser
- * composio login --api-key uak_xxx --org-id org-id --project-id proj-id
+ * composio login -y
  * ```
  */
-export const loginCmd = Command.make(
-  'login',
-  { noBrowser, apiKey: apiKeyOpt, orgId: orgIdOpt, projectId: projectIdOpt },
-  ({ noBrowser, apiKey, orgId, projectId }) =>
-    Effect.gen(function* () {
-      const ui = yield* TerminalUI;
-      const ctx = yield* ComposioUserContext;
+export const loginCmd = Command.make('login', { noBrowser, yes: yesOpt }, ({ noBrowser, yes }) =>
+  Effect.gen(function* () {
+    const ui = yield* TerminalUI;
+    const ctx = yield* ComposioUserContext;
 
-      yield* ui.intro('composio login');
+    yield* ui.intro('composio login');
 
-      // Non-interactive path: --api-key, --org-id, --project-id flags skip browser flow.
-      // All three must be provided together; partial flags are an error.
-      const nonInteractiveFlags = [apiKey, orgId, projectId];
-      const anyProvided = nonInteractiveFlags.some(Option.isSome);
-      const allProvided = nonInteractiveFlags.every(Option.isSome);
-
-      if (anyProvided && !allProvided) {
-        const missing = [
-          Option.isNone(apiKey) && '--api-key',
-          Option.isNone(orgId) && '--org-id',
-          Option.isNone(projectId) && '--project-id',
-        ].filter(Boolean);
-        yield* ui.log.error(`Missing required flag(s): ${missing.join(', ')}`);
-        yield* ui.log.info(
-          'Non-interactive login requires all three: --api-key, --org-id, --project-id'
+    if (ctx.isLoggedIn()) {
+      // Allow re-login when orgId/projectId are not yet set (old CLI login without multi-project support)
+      if (Option.isSome(ctx.data.orgId) && Option.isSome(ctx.data.projectId)) {
+        yield* ui.log.warn(`You're already logged in!`);
+        yield* ui.outro(
+          'If you want to log in with a different account, please run `composio logout` first.'
         );
-        yield* ui.outro('');
         return;
       }
+      yield* ui.log.step('Re-authenticating for multi-project support...');
+    }
 
-      // Strict verification: 400/401/403 from session/info are hard failures since
-      // the user explicitly provided the IDs.
-      if (
-        allProvided &&
-        Option.isSome(apiKey) &&
-        Option.isSome(orgId) &&
-        Option.isSome(projectId)
-      ) {
-        yield* Effect.logDebug('Non-interactive login with provided credentials');
-        yield* storeCredentials({
-          baseURL: ctx.data.baseURL,
-          uakApiKey: apiKey.value,
-          initialOrgId: orgId.value,
-          initialProjectId: projectId.value,
-          fallbackEmail: '',
-          strictVerification: true,
-        });
-        return;
-      }
-
-      if (ctx.isLoggedIn()) {
-        // Allow re-login when orgId/projectId are not yet set (old CLI login without multi-project support)
-        if (Option.isSome(ctx.data.orgId) && Option.isSome(ctx.data.projectId)) {
-          yield* ui.log.warn(`You're already logged in!`);
-          yield* ui.log.info(
-            `If you want to log in with a different account, please run \`composio logout\` first.`
-          );
-          yield* ui.outro('');
-          return;
-        }
-        yield* ui.log.step('Re-authenticating for multi-project support...');
-      }
-
-      yield* browserLogin({ scope: 'user', noBrowser });
-    })
+    yield* browserLogin({ scope: 'user', noBrowser, skipOrgProjectPicker: yes });
+  })
 ).pipe(Command.withDescription('Log in to the Composio SDK.'));
